+++ examples/fight/attribute/attribute.go
@@ -2,5 +2,5 @@ package attribute
 
 type Attribute struct {
 	AttrType AttrType
-	Value    float32
+	Value    int32
 }
